Moved test code to separate test file.
